feat(wasm): Add wasm support, update cryptovec
