+++ cryptovec/Cargo.toml
@@ -15,3 +15,6 @@ libc = "0.2"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 winapi = {version = "0.3", features = ["basetsd", "minwindef", "memoryapi"]}
+
+[dev-dependencies]
+wasm-bindgen-test = "0.3"
