+++ cryptovec/src/platform/mod.rs
@@ -1,0 +1,41 @@
+#[cfg(windows)]
+mod windows;
+
+#[cfg(not(windows))]
+#[cfg(not(target_arch = "wasm32"))]
+mod unix;
+
+#[cfg(target_arch = "wasm32")]
+mod wasm;
+
+// Re-export functions based on the platform
+#[cfg(windows)]
+pub use windows::{munlock,mlock,memset};
+
+#[cfg(not(windows))]
+#[cfg(not(target_arch = "wasm32"))]
+pub use unix::{munlock, mlock, memset};
+
+#[cfg(target_arch = "wasm32")]
+pub use wasm::{munlock, mlock, memset};
+
+#[cfg(test)]
+mod tests {
+    use wasm_bindgen_test::wasm_bindgen_test;
+
+    use super::*;
+
+    #[wasm_bindgen_test]
+    fn test_memset() {
+        let mut buf = vec![0u8; 10];
+        memset(buf.as_mut_ptr(), 0xff, buf.len());
+        assert_eq!(buf, vec![0xff; 10]);
+    }
+
+    #[wasm_bindgen_test]
+    fn test_memset_partial() {
+        let mut buf = vec![0u8; 10];
+        memset(buf.as_mut_ptr(), 0xff, 5);
+        assert_eq!(buf, [0xff, 0xff, 0xff, 0xff, 0xff, 0, 0, 0, 0, 0]);
+    }
+}
